当当全局抓取；根据isbn抓取豆瓣，亚马逊；meta表字段优化；图书查询添加查询字段

+++ src/main/java/com/apabi/flow/jd/model/JdMetadata.java
@@ -9,7 +9,7 @@ import java.util.Date;
 
 /**
  * @Author pipi
- * @Date 2018-11-30 17:34
+ * @Date 2018-12-04 10:56
  **/
 @Data
 @NoArgsConstructor
@@ -30,7 +30,7 @@ public class JdMetadata {
 
     private String binding;
 
-    private Short pages;
+    private Integer pages;
 
     private String language;
